Update main.py
Now it can take multiple inputs (simply enter empty line to stop). Also made it so that the song stops after like 8 seconds. But turns out this doesn't work for any song not in English, which is quite the problem for my specific case. Also made everything a little bit faster.